Checkpoint the Oct2025 change (#1)
* update dtype argument

* update default

* update training with pretokenized dataset fomr hf

* add contratints to the max length when using pretokenized dataset

* integrate neuclir jsonl encoding

* update for pretokenized (but maybe is not needed anymore)

+++ src/tevatron/retriever/arguments.py
@@ -64,7 +64,7 @@ class ModelArguments:
 
     # for Jax training
     dtype: Optional[str] = field(
-        default="bfloat16",
+        default="float16",
         metadata={
             "help": "Floating-point format in which the model weights should be initialized and trained. Choose one "
                     "of `[float32, float16, bfloat16]`. "
